adding dry-run and verbose options

+++ mackup/application.py
@@ -12,9 +12,7 @@ from . import utils
 
 class ApplicationProfile(object):
 
-    """Instantiate this class with application specific data."""
-
-    def __init__(self, mackup, files):
+    def __init__(self, mackup, files, dry_run, verbose):
         """
         Create an ApplicationProfile instance.
 
@@ -27,6 +25,22 @@ class ApplicationProfile(object):
 
         self.mackup = mackup
         self.files = list(files)
+        self.dry_run = dry_run
+        self.verbose = verbose
+
+
+    def getFilepaths(self, filename):
+        """
+        Get home and mackup filepaths for given file
+
+        Args:
+            filepath (str)
+
+        Returns:
+            home_filepath, mackup_filepath (str, str)
+        """
+        return (os.path.join(os.environ['HOME'], filename),
+                os.path.join(self.mackup.mackup_folder, filename))
 
     def backup(self):
         """
@@ -47,18 +61,21 @@ class ApplicationProfile(object):
         """
         # For each file used by the application
         for filename in self.files:
-            # Get the full path of each file
-            filepath = os.path.join(os.environ['HOME'], filename)
-            mackup_filepath = os.path.join(self.mackup.mackup_folder, filename)
+            (home_filepath, mackup_filepath) = self.getFilepaths(filename)
 
             # If the file exists and is not already a link pointing to Mackup
-            if ((os.path.isfile(filepath) or os.path.isdir(filepath))
-                and not (os.path.islink(filepath)
+            if ((os.path.isfile(home_filepath) or os.path.isdir(home_filepath))
+                and not (os.path.islink(home_filepath)
                          and (os.path.isfile(mackup_filepath)
                               or os.path.isdir(mackup_filepath))
-                         and os.path.samefile(filepath, mackup_filepath))):
+                         and os.path.samefile(home_filepath, mackup_filepath))):
 
-                print("Backing up {} ...".format(filepath))
+                print("Backing up {} ...".format("" if self.verbose else filename))
+                if self.verbose:
+                    print("  {}\n  to\n  {}".format(home_filepath, mackup_filepath))
+
+                if self.dry_run:
+                    continue
 
                 # Check if we already have a backup
                 if os.path.exists(mackup_filepath):
@@ -82,18 +99,23 @@ class ApplicationProfile(object):
                         # Delete the file in Mackup
                         utils.delete(mackup_filepath)
                         # Copy the file
-                        utils.copy(filepath, mackup_filepath)
+                        utils.copy(home_filepath, mackup_filepath)
                         # Delete the file in the home
-                        utils.delete(filepath)
+                        utils.delete(home_filepath)
                         # Link the backuped file to its original place
-                        utils.link(mackup_filepath, filepath)
+                        utils.link(mackup_filepath, home_filepath)
                 else:
                     # Copy the file
-                    utils.copy(filepath, mackup_filepath)
+                    utils.copy(home_filepath, mackup_filepath)
                     # Delete the file in the home
-                    utils.delete(filepath)
+                    utils.delete(home_filepath)
                     # Link the backuped file to its original place
-                    utils.link(mackup_filepath, filepath)
+                    utils.link(mackup_filepath, home_filepath)
+            elif self.verbose:
+              print("Doing nothing\n  {}\n  {}".format(
+                  home_filepath, 
+                  "is already backed up to\n  " + mackup_filepath
+                      if os.path.exists(home_filepath) else "does not exist"))
 
     def restore(self):
         """
@@ -111,9 +133,7 @@ class ApplicationProfile(object):
         """
         # For each file used by the application
         for filename in self.files:
-            # Get the full path of each file
-            mackup_filepath = os.path.join(self.mackup.mackup_folder, filename)
-            home_filepath = os.path.join(os.environ['HOME'], filename)
+            (home_filepath, mackup_filepath) = self.getFilepaths(filename)
 
             # If the file exists and is not already pointing to the mackup file
             # and the folder makes sense on the current platform (Don't sync
@@ -126,7 +146,12 @@ class ApplicationProfile(object):
             supported = utils.can_file_be_synced_on_current_platform(filename)
 
             if file_or_dir_exists and not pointing_to_mackup and supported:
-                print("Restoring {} ...".format(home_filepath))
+                print("Restoring {} ...".format("" if self.verbose else filename))
+                if self.verbose:
+                    print("  linking {}\n  to      {}".format(home_filepath, mackup_filepath))
+
+                if self.dry_run:
+                    continue
 
                 # Check if there is already a file in the home folder
                 if os.path.exists(home_filepath):
@@ -149,6 +174,11 @@ class ApplicationProfile(object):
                         utils.link(mackup_filepath, home_filepath)
                 else:
                     utils.link(mackup_filepath, home_filepath)
+            elif self.verbose:
+                print("Doing nothing\n  {}\n  {}".format(
+                    mackup_filepath,
+                    "already linked by\n  " + home_filepath
+                        if os.path.exists(home_filepath) else "does not exist"))
 
     def uninstall(self):
         """
@@ -167,19 +197,25 @@ class ApplicationProfile(object):
         """
         # For each file used by the application
         for filename in self.files:
-            # Get the full path of each file
-            mackup_filepath = os.path.join(self.mackup.mackup_folder, filename)
-            home_filepath = os.path.join(os.environ['HOME'], filename)
+            (home_filepath, mackup_filepath) = self.getFilepaths(filename)
 
             # If the mackup file exists
             if (os.path.isfile(mackup_filepath)
                     or os.path.isdir(mackup_filepath)):
                 # Check if there is a corresponding file in the home folder
                 if os.path.exists(home_filepath):
-                    print("Reverting {} ...".format(home_filepath))
+                    print("Reverting {} ...".format(mackup_filepath if self.verbose else filename))
+                    if self.verbose:
+                        print("  at {}".format(home_filepath))
+
+                    if self.dry_run:
+                        continue
+
                     # If there is, delete it as we are gonna copy the Dropbox
                     # one there
                     utils.delete(home_filepath)
 
                     # Copy the Dropbox file to the home folder
                     utils.copy(mackup_filepath, home_filepath)
+            elif self.verbose:
+                print("Doing nothing, {} does not exist".format(mackup_filepath))
